survey launch button hidden for archive
